added unittest to yaccer, lexer and query.
First complete version. 
Subsequent queries can be of any depth and, newly, an empty subsequent query indicates a reduce.
A reduce is the collapse of a dictionary of DataFrames to a single DataFrame.
Also newly, the query result dictionary _key is available at the query prompt and that is handy for reduced query groups.